Rename ProjectBuilder to ProjectBuilder2015 and ProjectBuilder2017
