+++ src/ApiPort.VisualStudio.2015/ApiPort.VisualStudio.2015.csproj
@@ -137,7 +137,7 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ProjectBuilder.cs" />
+    <Compile Include="ProjectBuilder2015.cs" />
     <Compile Include="ServiceProvider.cs" />
     <Compile Include="VSThreadingService.cs" />
   </ItemGroup>
