+++ src/ApiPort.VisualStudio.2017/ServiceProvider.cs
@@ -21,7 +21,7 @@ namespace ApiPortVS.VS2017
             builder.RegisterType<VSThreadingService>()
                 .As<IVSThreadingService>()
                 .SingleInstance();
-            builder.RegisterType<ProjectBuilder>()
+            builder.RegisterType<ProjectBuilder2015>()
                 .As<IProjectBuilder>()
                 .SingleInstance();
